Split Response class
