+++ lib/PuppeteerSharp/Response.cs
@@ -1,122 +1,60 @@
-using System;
 using System.Collections.Generic;
 using System.Net;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
 using PuppeteerSharp.Helpers.Json;
-using PuppeteerSharp.Messaging;
 
 namespace PuppeteerSharp
 {
     /// <inheritdoc/>
-    public class Response : IResponse
+    public abstract class Response : IResponse
     {
-        private static readonly Regex _extraInfoLines = new(@"[^ ]* [^ ]* (?<text>.*)", RegexOptions.Multiline);
-        private readonly CDPSession _client;
-        private readonly bool _fromDiskCache;
-        private byte[] _buffer;
-
-        internal Response(
-            CDPSession client,
-            Request request,
-            ResponsePayload responseMessage,
-            ResponseReceivedExtraInfoResponse extraInfo)
+        internal Response()
         {
-            _client = client;
-            Request = request;
-            Status = extraInfo != null ? extraInfo.StatusCode : responseMessage.Status;
-            StatusText = ParseStatusTextFromExtrInfo(extraInfo) ?? responseMessage.StatusText;
-            Url = request.Url;
-            _fromDiskCache = responseMessage.FromDiskCache;
-            FromServiceWorker = responseMessage.FromServiceWorker;
-
-            Headers = new Dictionary<string, string>(StringComparer.InvariantCultureIgnoreCase);
-            var headers = extraInfo != null ? extraInfo.Headers : responseMessage.Headers;
-            if (headers != null)
-            {
-                foreach (var keyValue in headers)
-                {
-                    Headers[keyValue.Key] = keyValue.Value;
-                }
-            }
-
-            SecurityDetails = responseMessage.SecurityDetails;
-            RemoteAddress = new RemoteAddress
-            {
-                IP = responseMessage.RemoteIPAddress,
-                Port = responseMessage.RemotePort,
-            };
         }
 
         /// <inheritdoc/>
-        public string Url { get; }
+        public RemoteAddress RemoteAddress { get; protected init; }
 
         /// <inheritdoc/>
-        public Dictionary<string, string> Headers { get; }
-
-        /// <inheritdoc/>
-        public HttpStatusCode Status { get; }
+        public string Url { get; protected init; }
 
         /// <inheritdoc/>
         public bool Ok => Status == 0 || ((int)Status >= 200 && (int)Status <= 299);
 
-        /// <inheritdoc cref="Request"/>
-        public Request Request { get; }
+        /// <inheritdoc/>
+        public HttpStatusCode Status { get; protected init; }
+
+        /// <inheritdoc/>
+        public string StatusText { get; protected init; }
+
+        /// <inheritdoc/>
+        public Dictionary<string, string> Headers { get; protected init; }
 
         /// <inheritdoc/>
         IRequest IResponse.Request => Request;
 
         /// <inheritdoc/>
-        public bool FromCache => _fromDiskCache || (Request?.FromMemoryCache ?? false);
+        public abstract bool FromCache { get; }
 
         /// <inheritdoc/>
-        public SecurityDetails SecurityDetails { get; }
+        public SecurityDetails SecurityDetails { get; protected init; }
 
         /// <inheritdoc/>
-        public bool FromServiceWorker { get; }
-
-        /// <inheritdoc/>
-        public string StatusText { get; }
-
-        /// <inheritdoc/>
-        public RemoteAddress RemoteAddress { get; }
+        public bool FromServiceWorker { get; protected init; }
 
         /// <inheritdoc/>
         public IFrame Frame => Request.Frame;
 
-        internal TaskCompletionSource<bool> BodyLoadedTaskWrapper { get; } = new(TaskCreationOptions.RunContinuationsAsynchronously);
+        /// <inheritdoc cref="Request"/>
+        protected Request Request { get; init; }
 
         /// <summary>
         /// Returns a Task which resolves to a buffer with response body.
         /// </summary>
         /// <returns>A Task which resolves to a buffer with response body.</returns>
-        public async ValueTask<byte[]> BufferAsync()
-        {
-            if (_buffer == null)
-            {
-                await BodyLoadedTaskWrapper.Task.ConfigureAwait(false);
-
-                try
-                {
-                    var response = await _client.SendAsync<NetworkGetResponseBodyResponse>("Network.getResponseBody", new NetworkGetResponseBodyRequest
-                    {
-                        RequestId = Request.RequestId,
-                    }).ConfigureAwait(false);
-
-                    _buffer = response.Base64Encoded
-                        ? Convert.FromBase64String(response.Body)
-                        : Encoding.UTF8.GetBytes(response.Body);
-                }
-                catch (Exception ex)
-                {
-                    throw new BufferException("Unable to get response body", ex);
-                }
-            }
-
-            return _buffer;
-        }
+        public abstract ValueTask<byte[]> BufferAsync();
 
         /// <inheritdoc/>
         public async Task<string> TextAsync() => Encoding.UTF8.GetString(await BufferAsync().ConfigureAwait(false));
@@ -126,30 +64,5 @@ namespace PuppeteerSharp
 
         /// <inheritdoc/>
         public async Task<T> JsonAsync<T>() => (await JsonAsync().ConfigureAwait(false)).ToObject<T>(true);
-
-        private string ParseStatusTextFromExtrInfo(ResponseReceivedExtraInfoResponse extraInfo)
-        {
-            if (extraInfo == null || extraInfo.HeadersText == null)
-            {
-                return null;
-            }
-
-            var lines = extraInfo.HeadersText.Split('\r');
-            if (lines.Length == 0)
-            {
-                return null;
-            }
-
-            var firstLine = lines[0];
-
-            var match = _extraInfoLines.Match(firstLine);
-            if (!match.Success)
-            {
-                return null;
-            }
-
-            var statusText = match.Groups["text"].Value;
-            return statusText;
-        }
     }
 }
