+++ lib/PuppeteerSharp/WebWorker.cs
@@ -1,9 +1,5 @@
-using System;
-using System.Linq;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
-using PuppeteerSharp.Helpers.Json;
-using PuppeteerSharp.Messaging;
 
 namespace PuppeteerSharp
 {
